hub-depot-climb auto that doesn't really work but we're calling it for tonight

+++ src/main/java/frc/robot/Constants.java
@@ -363,7 +363,7 @@ public final class Constants {
     public static final boolean mirrorForRedAlliance = true;
 
     // Seconds left when robot tries to climb
-    public static final int climbTime = 5;
+    public static final double climbTime = 5;
 
     public static double autoLinearKp = 10.0;
     public static double autoLinearKi = 0.0;
